move sync to header
