+++ client/src/app/auth/header/sync-account-button.tsx
